requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Neuroinformatics
navis>=1.0.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Data processing
openpyxl>=3.0.0
pyyaml>=6.0

# Progress bars and utilities
tqdm>=4.60.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional dependencies for analysis
scikit-learn>=1.0.0
seaborn>=0.11.0

# File I/O and path handling
pathlib2>=2.3.0

# Statistical analysis
statsmodels>=0.13.0 